# specs/001-code-learning-platform/contracts/api-spec.yaml
openapi: 3.0.3
info:
  title: AI Code Learning Platform API
  description: |
    REST API for the AI Code Learning Platform - Phase 1 MVP.

    This platform transforms AI-generated code into beginner-friendly learning materials
    through task-based code uploads, automatic analysis, document generation, and Q&A.

    Powered by Google Gemini 2.5 Flash for fast, cost-effective AI content generation.

    **Authentication**: JWT-based with HTTPOnly cookies
    **Base URL**: `/api/v1`
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:8000/api/v1
    description: Development server
  - url: https://api.codelearning.com/api/v1
    description: Production server
...
paths:
...
  /projects/{project_id}/tasks:
    get:
      tags: [Tasks]
      summary: List tasks in project
      description: Get all tasks in sequential order
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Tasks retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  tasks:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'
                  total:
                    type: integer

    post:
      tags: [Tasks]
      summary: Create new task
      description: Create task and upload code (auto-assigns sequential task number)
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [title, upload_method]
              properties:
                title:
                  type: string
                  minLength: 5
                  maxLength: 255
                  example: Addition Function
                description:
                  type: string
                  maxLength: 500
                  example: Building a basic addition function
                upload_method:
                  type: string
                  enum: [file, folder, paste]
                  example: file
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                  maxItems: 20
                  description: Files to upload (for file/folder method)
                code_text:
                  type: string
                  description: Pasted code (for paste method)
                language:
                  type: string
                  description: Programming language (required for paste method)
                  example: python
      responses:
        '201':
          description: Task created and code uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid input or file validation failed
        '413':
          description: Upload size exceeds 10MB limit
  /tasks/{task_id}:
    get:
      tags: [Tasks]
      summary: Get task details
      description: Get task with all associated data (code, document, practice, progress)
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: Task retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDetail'

    patch:
      tags: [Tasks]
      summary: Update task
      description: Update task title or description
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
      responses:
        '200':
          description: Task updated

    delete:
      tags: [Tasks]
      summary: Soft delete task
      description: Move task to trash (30-day retention)
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '204':
          description: Task moved to trash
  /tasks/{task_id}/code:
    get:
      tags: [Tasks]
      summary: Get uploaded code
      description: Get code files and metadata for a task
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: Code retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadedCode'

  # ==================== Documents ====================
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: JWT access token in HTTPOnly cookie
  parameters:
    ProjectId:
      name: project_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    TaskId:
      name: task_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        task_number:
          type: integer
          example: 1
        title:
          type: string
        description:
          type: string
          nullable: true
        upload_method:
          type: string
          enum: [file, folder, paste]
        created_at:
          type: string
          format: date-time
        deletion_status:
          type: string
          enum: [active, trashed]
        document_status:
          type: string
          enum: [pending, in_progress, completed, failed]
          description: Learning document generation status
    TaskDetail:
      allOf:
        - $ref: '#/components/schemas/Task'
        - type: object
          properties:
            code:
              $ref: '#/components/schemas/UploadedCode'
            document:
              $ref: '#/components/schemas/LearningDocument'
              nullable: true
            practice_problems:
              type: array
              items:
                $ref: '#/components/schemas/PracticeProblem'
            progress:
              $ref: '#/components/schemas/Progress'
    UploadedCode:
      type: object
      properties:
        detected_language:
          type: string
          example: python
        complexity_level:
          type: string
          enum: [beginner, intermediate, advanced]
        total_lines:
          type: integer
        total_files:
          type: integer
        files:
          type: array
          items:
            $ref: '#/components/schemas/CodeFile'
    CodeFile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        file_name:
          type: string
          example: calculator.py
        file_path:
          type: string
          nullable: true
          example: src/calculator.py
        file_size_bytes:
          type: integer
        content_url:
          type: string
          description: URL to download file content
          example: /api/v1/files/abc123/content
    LearningDocument:
      type: object
      properties:
        id:
          type: string
          format: uuid
        generation_status:
          type: string
          enum: [pending, in_progress, completed, failed]
        content:
          type: object
          description: 7-chapter structured content
          properties:
            chapter1:
              type: object
              properties:
                title:
                  type: string
                summary:
                  type: string
            chapter2:
              type: object
              properties:
                title:
                  type: string
                concepts:
                  type: array
                  items:
                    type: object
            # ... other chapters
        generated_at:
          type: string
          format: date-time
    PracticeProblem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        problem_number:
          type: integer
          minimum: 1
          maximum: 5
        problem_type:
          type: string
          enum: [follow_along, fill_blanks, modify, debug, create]
        problem_statement:
          type: string
        learning_objective:
          type: string
        hints:
          type: array
          items:
            type: string
          description: Progressive hints (reveal one at a time)
        model_solution:
          type: string
          description: Complete solution (hidden until requested)
        expected_output:
          type: string
...
    Progress:
      type: object
      properties:
        chapters_completed:
          type: array
          items:
            type: integer
          example: [1, 2, 3]
        document_read_percentage:
          type: integer
          minimum: 0
          maximum: 100
        problems_completed:
          type: array
          items:
            type: integer
          example: [1, 2, 3]
        practice_completion_count:
          type: integer
          minimum: 0
          maximum: 5
        questions_asked_count:
          type: integer
        task_completed:
          type: boolean
        task_completion_date:
          type: string
          format: date-time
          nullable: true
...